requirements.txt
# Django Framework
Django>=6.0,<7.0

# Image processing library (recommandé pour Django, même avec FileField)
Pillow>=10.0.0

# Interface admin améliorée
django-admin-interface>=3.2.0
django-colorfield>=0.9.0

# Database (SQLite est inclus avec Python, mais on peut ajouter d'autres backends si nécessaire)
# Pour PostgreSQL: psycopg2-binary>=2.9.0
# Pour MySQL: mysqlclient>=2.2.0

# Génération de PDFs
reportlab>=4.0.0

